Refactor CLI runtime into SRP modules
